requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
clams-python==1.1.2
click==8.1.7
deepdiff==6.7.1
Flask==3.0.2
Flask-RESTful==0.3.10
gunicorn==21.2.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lapps==0.0.2
lxml==5.1.0
MarkupSafe==2.1.5
mmif-python==1.0.9
ordered-set==4.1.0
packaging==23.2
platformdirs==4.2.0
pydantic==1.10.14
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.18.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zeep==4.2.1
